# auction-service/src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "Swagger Demo Project",
    "description": "Implementation of Swagger with TypeScript"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": ""
    }
  ],
  "paths": {
    "/checkHealth": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auctions/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auctions/{auctionId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auctions/{auctionId}/flag": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "auctionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/categories/categories/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/watchlists/watchlists/": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}